src/gridsingularity/apps/simulation/serializers.py
from rest_framework import serializers
from .models import SimulationResult

class SimulationResultSerializer(serializers.ModelSerializer):
    class Meta:
        model = SimulationResult
        fields = ('id', 'active', 'reactive')
