requirements/base.txt
Django==3.0.6
pandapower==2.2.2
djangorestframework==3.11.0